fix(spi): rx slot freeing in Spi_NotifyRxData
The function is responsible for freeing the rx slot of a completed transaction. Before it poped the slot from the ring buffer before accessing the data. This can lead to a race condition where another transaction writes to this slot. Therefore, the pop is only done after accessing the rx data.