feat(fdcan): imporved timestamp creation and handling to make it more precise
The FDCAN ISR loops over all received frames and parses them. The timestamp was only sampled once and then assigned to all frames looped over. Now the timestamp is computed properly: get a global timestamp on entering the ISR and then using the hardsware timestamp of each frame to add them onto the global timestamp to compute the frames "absolute" timestamp. For this, also TIM3 as external timestamp source was set up for fdcan.